Added Functionality to Spontaneous Views
Added functionality to the Spontaneous Screen and the Spontaneous Active Screen. The Spontaneous Screen now loads and renders internal data. Selecting any rule on the Spontaneous Screen now loads the details of that specific rule. Other minor display updates. (The Spontaneou Screen's display has temporarily been changed to a ListView intead of a Grid until I learn how to make it a grid again. However, it meets the criteria for checkpoint 3)